Tower placement, ui changes, and collision
you can now only place towers on the green cube (floor) and towers can no longer be stacked on top of each other. when you click on a tower a unique canvas panel will show for the tower type to display tower data and currently implementing method to sell tower and gain reduced gold back as well as tower panels not showing while you are holding a tower to place.